refact: move workflow-related views out of views_ingest to their own file
